+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Unreleased
 
+- 2022-03-16 -- [@chasefleming](https://github.com/chasefleming): Warn about field renamings/deprecations. To turn on warnings, set config `log.level` to `2`.
+
+```js
+sdk.config("logger.level", 2)
+```
+
 - 2022-02-11 -- Uses Buffer from @onflow/rlp in encode.
 - 2022-02-11 -- Injects Buffer from @onflow/rlp to transport send modules.
 - 2022-02-04 -- [@chasefleming](https://github.com/chasefleming): Add options for for getting account by block height.